Move dynamically loading JSON logic into its own service. Replace try catch block for writing results with one liner.
